Move `BinaryPath` and `BashBinary` from `pants.engine.process` to `pants.core.util_rules.system_binaries`
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
