+++ src/python/pants/backend/python/util_rules/pex_environment.py
@@ -11,11 +11,9 @@ from typing import Mapping
 from pants.core.subsystems.python_bootstrap import PythonBootstrap
 from pants.core.util_rules import subprocess_environment, system_binaries
 from pants.core.util_rules.subprocess_environment import SubprocessEnvironmentVars
-from pants.core.util_rules.system_binaries import PythonBinary
-from pants.engine import process
+from pants.core.util_rules.system_binaries import BinaryPath, PythonBinary
 from pants.engine.engine_aware import EngineAwareReturnType
 from pants.engine.environment import Environment
-from pants.engine.process import BinaryPath
 from pants.engine.rules import collect_rules, rule
 from pants.option.global_options import NamedCachesDirOption
 from pants.option.option_types import BoolOption, IntOption, StrListOption
@@ -226,7 +224,6 @@ class CompletePexEnvironment:
 def rules():
     return [
         *collect_rules(),
-        *process.rules(),
         *subprocess_environment.rules(),
         *system_binaries.rules(),
     ]
