+++ src/python/pants/core/util_rules/system_binaries_test.py
@@ -6,11 +6,12 @@ from __future__ import annotations
 import os
 import sys
 from dataclasses import dataclass
+from pathlib import Path
 
 import pytest
 
 from pants.core.util_rules import system_binaries
-from pants.core.util_rules.system_binaries import PythonBinary
+from pants.core.util_rules.system_binaries import BinaryPathRequest, BinaryPaths, PythonBinary
 from pants.engine.internals.selectors import Get
 from pants.engine.process import Process, ProcessResult
 from pants.engine.rules import QueryRule, rule
@@ -37,8 +38,99 @@ async def python_binary_version(python_binary: PythonBinary) -> PythonBinaryVers
 @pytest.fixture
 def rule_runner() -> RuleRunner:
     return RuleRunner(
-        rules=[*system_binaries.rules(), python_binary_version, QueryRule(PythonBinaryVersion, [])]
+        rules=[
+            *system_binaries.rules(),
+            python_binary_version,
+            QueryRule(PythonBinaryVersion, []),
+            QueryRule(BinaryPaths, [BinaryPathRequest]),
+        ]
     )
+
+
+def test_find_binary_non_existent(rule_runner: RuleRunner, tmp_path: Path) -> None:
+    binary_paths = rule_runner.request(
+        BinaryPaths, [BinaryPathRequest(binary_name="nonexistent-bin", search_path=[str(tmp_path)])]
+    )
+    assert binary_paths.first_path is None
+
+
+class MyBin:
+    binary_name = "mybin"
+
+    @classmethod
+    def create(cls, directory: Path) -> Path:
+        directory.mkdir(parents=True, exist_ok=True)
+        exe = directory / cls.binary_name
+        exe.touch(mode=0o755)
+        return exe
+
+
+def test_find_binary_on_path_without_bash(rule_runner: RuleRunner, tmp_path: Path) -> None:
+    # Test that locating a binary on a PATH which does not include bash works (by recursing to
+    # locate bash first).
+    binary_dir_abs = tmp_path / "bin"
+    binary_path_abs = MyBin.create(binary_dir_abs)
+
+    binary_paths = rule_runner.request(
+        BinaryPaths,
+        [BinaryPathRequest(binary_name=MyBin.binary_name, search_path=[str(binary_dir_abs)])],
+    )
+    assert binary_paths.first_path is not None
+    assert binary_paths.first_path.path == str(binary_path_abs)
+
+
+def test_find_binary_file_path(rule_runner: RuleRunner, tmp_path: Path) -> None:
+    binary_path_abs = MyBin.create(tmp_path)
+
+    binary_paths = rule_runner.request(
+        BinaryPaths,
+        [
+            BinaryPathRequest(
+                binary_name=MyBin.binary_name,
+                search_path=[str(binary_path_abs)],
+            )
+        ],
+    )
+    assert binary_paths.first_path is None, "By default, PATH file entries should not be checked."
+
+    binary_paths = rule_runner.request(
+        BinaryPaths,
+        [
+            BinaryPathRequest(
+                binary_name=MyBin.binary_name,
+                search_path=[str(binary_path_abs)],
+                check_file_entries=True,
+            )
+        ],
+    )
+    assert binary_paths.first_path is not None
+    assert binary_paths.first_path.path == str(binary_path_abs)
+
+
+def test_find_binary_respects_search_path_order(rule_runner: RuleRunner, tmp_path: Path) -> None:
+    binary_path_abs1 = MyBin.create(tmp_path / "bin1")
+    binary_path_abs2 = MyBin.create(tmp_path / "bin2")
+    binary_path_abs3 = MyBin.create(tmp_path / "bin3")
+
+    binary_paths = rule_runner.request(
+        BinaryPaths,
+        [
+            BinaryPathRequest(
+                binary_name=MyBin.binary_name,
+                search_path=[
+                    str(binary_path_abs1.parent),
+                    str(binary_path_abs2),
+                    str(binary_path_abs3.parent),
+                ],
+                check_file_entries=True,
+            )
+        ],
+    )
+    assert binary_paths.first_path is not None
+    assert binary_paths.first_path.path == str(binary_path_abs1)
+    assert [str(p) for p in (binary_path_abs1, binary_path_abs2, binary_path_abs3)] == [
+        binary_path.path for binary_path in binary_paths.paths
+    ]
 
 
 def test_python_binary(rule_runner: RuleRunner) -> None:
@@ -47,7 +139,7 @@ def test_python_binary(rule_runner: RuleRunner) -> None:
     assert python_binary_version.version.startswith("Python 3.")
 
 
-def test_interpreter_search_path_file_entries() -> None:
+def test_python_interpreter_search_path_file_entries() -> None:
     rule_runner = RuleRunner(
         rules=[*system_binaries.rules(), QueryRule(PythonBinary, input_types=())]
     )
